+++ mautrix/crypto/__init__.py
@@ -1,7 +1,6 @@
 from .account import OlmAccount
 from .key_share import RejectKeyShare
 from .sessions import InboundGroupSession, OutboundGroupSession, Session
-from .types import DecryptedOlmEvent, DeviceIdentity, TrustState
 
 # These have to be last
 from .store import (  # isort: skip